fix(DPLAN-18071): fix unresolved import after helper rename
redirectToStatementListWithResolvedToast was renamed to
redirectToStatementListWithGroupResolvedToast but the file and one of
the two call sites weren't updated to match, breaking the import.
