+++ client/js/components/procedure/Shared/utils/redirectToStatementListWithGroupResolvedToast.js
