+++ client/js/components/statement/listStatements/ListStatements.vue
@@ -683,13 +683,13 @@ export default {
 
     // Show the "group resolved" toast once when arriving here after the last group member was detached
     notifyClusterResolved () {
-      const ToastData = lscache.get(`${this.procedureId}:clusterElementDetached`)
+      const detachedToastData = lscache.get(`${this.procedureId}:clusterElementDetached`)
 
-      if (ToastData) {
+      if (detachedToastData) {
         // One-shot flag: remove it so a reload does not re-trigger the toast
         lscache.remove(`${this.procedureId}:clusterElementDetached`)
 
-        const { statementId, clusterId: detachedClusterId } = JSON.parse(ToastData)
+        const { statementId, clusterId: detachedClusterId } = JSON.parse(detachedToastData)
 
         dplan.notify.notify('confirm', Translator.trans('confirm.statement.detach.cluster.element', {
           statementId,
